+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - fix switch network issues
 
 ### Improvements
+- Make 'Local test net' optional
 
 # v0.4.1 (2019.02.12)